+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `contains` to StringPredicate
 - Add `clearLocalStorage` to SeleniumEnvironment
 - Fix `selectByLabel` and `selectByValue` of various SelectComponents
+- ClickableSeleniumComponent should respect the short timeout
 
 ## selenium-components-0.4.2